Issue 43896: Clean up JdbcMetaDataLocator hierarchy
