+++ api/src/org/labkey/api/data/SchemaNameCache.java
@@ -15,9 +15,7 @@
  */
 package org.labkey.api.data;
 
-import org.jetbrains.annotations.Nullable;
 import org.labkey.api.cache.BlockingCache;
-import org.labkey.api.cache.CacheLoader;
 import org.labkey.api.cache.CacheManager;
 import org.labkey.api.collections.CaseInsensitiveTreeMap;
 import org.labkey.api.data.dialect.JdbcMetaDataLocator;
