+++ bigiron/src/org/labkey/bigiron/sas/SasDialect.java
@@ -26,10 +26,10 @@ import org.labkey.api.data.RowTrackingResultSetWrapper;
 import org.labkey.api.data.RuntimeSQLException;
 import org.labkey.api.data.SQLFragment;
 import org.labkey.api.data.SqlExecutor;
+import org.labkey.api.data.dialect.StandardJdbcMetaDataLocator;
 import org.labkey.api.data.dialect.JdbcHelper;
 import org.labkey.api.data.dialect.JdbcMetaDataLocator;
 import org.labkey.api.data.dialect.SimpleSqlDialect;
-import org.labkey.api.data.dialect.StandardJdbcMetaDataLocator;
 import org.labkey.api.data.dialect.StandardTableResolver;
 import org.labkey.api.data.dialect.StatementWrapper;
 import org.labkey.api.data.dialect.TableResolver;
